@@ -13,8 +13,8 @@ namespace DashboardAPI.Controllers
     [Route(".well-known")]
     public class FreeSSLController : Controller
     {
-        private readonly IHostingEnvironment _env;
-        public FreeSSLController(IHostingEnvironment env)
+        private readonly IWebHostEnvironment _env;
+        public FreeSSLController(IWebHostEnvironment env)
         {
             _env = env;
         }
