@@ -1,5 +1,4 @@
 ﻿using Confluent.Kafka;
-using Confluent.Kafka.Serialization;
 using Microsoft.Extensions.Options;
 using Sticky.Models.Druid;
 using Sticky.Models.Etc;
@@ -14,31 +13,26 @@ namespace Sticky.Repositories.Script.Implementions
     public class KafkaClient : IKafkaClient
     {
         private readonly IRedisCache _redisCache;
-        private readonly Producer<Null, string> _producer;
+        private readonly IProducer<Null, string> _producer;
         public KafkaClient(IOptions<ScriptAPISetting> configuratoin, IRedisCache redisCache)
         {
             _redisCache = redisCache;
             var kafkaaddresses = configuratoin.Value.KafkaAddress;
-            var config = new Dictionary<string, object>
-            {
-                { "bootstrap.servers", kafkaaddresses}
-            };
+            var config = new ProducerConfig { BootstrapServers = kafkaaddresses };
 
-            _producer = new Producer<Null, string>(config, null, new StringSerializer(Encoding.UTF8));
+            _producer = new ProducerBuilder<Null, string>(config).Build();
         }
         public KafkaClient(IRedisCache redisCache)
         {
             _redisCache = redisCache;
-            var config = new Dictionary<string, object>
-      {
-        { "bootstrap.servers", "<##Change:KafkaAddress##>"}
-      };
-            _producer = new Producer<Null, string>(config, null, new StringSerializer(Encoding.UTF8));
+            var config = new ProducerConfig { BootstrapServers = "<##Change:KafkaAddress##>" };
+
+            _producer = new ProducerBuilder<Null, string>(config).Build();
         }
         public async Task SendMessage(DruidData message)
         {
 
-         await _producer.ProduceAsync("Sticky", null, Newtonsoft.Json.JsonConvert.SerializeObject(message));
+         await _producer.ProduceAsync("Sticky",new Message<Null, string> { Value = Newtonsoft.Json.JsonConvert.SerializeObject(message) });
         }
 
 
